refactor: fix rental detail frame
